remove extra query for claims button
combine it with the query for the overview menu instead

also pull some of the components that were in MainHeader.tsx into their own files
